# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Obiara API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Authoritative REST contract for Obiara member clients. Durable API changes
    are additive within v1. Retriable writes require an Idempotency-Key.
...
servers:
  - url: /
    description: API root
...
paths:
...
  /v1/matchmakers:
    get:
      operationId: listLicensedMatchmakers
      tags: [Commerce]
      summary: List currently licensed matchmakers
      security:
        - MemberBearer: []
      responses:
        "200":
          description: Current, privacy-bounded licensed profiles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchmakerProfileListEnvelope"
        "401":
          description: A valid member session is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          $ref: "#/components/responses/InternalError"
  /v1/matchmaker-engagements:
    get:
      operationId: listMemberMatchmakerEngagements
      tags: [Commerce]
      summary: List the authenticated member's matchmaker engagements
      security:
        - MemberBearer: []
      responses:
        "200":
          description: Member-owned engagements.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchmakerEngagementListEnvelope"
        "401":
          description: A valid member session is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: bookMatchmakerEngagement
      tags: [Commerce]
      summary: Book immutable terms with a currently licensed matchmaker
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchmakerBookingInput"
      responses:
        "201":
          description: Engagement booked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchmakerEngagementEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          description: A valid member session is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Idempotency or concurrent-write conflict.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "503":
          $ref: "#/components/responses/InternalError"
...
  /v1/circles/{circleId}/oware:
    post:
      operationId: createOwareGame
      tags: [Community]
      summary: Start a private Oware game in a two-member circle
      description: |
        The server derives both players from current circle membership. The
        request is rejected unless exactly two active members are present.
        Privacy-keyed player and room references are never projected.
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "201":
          description: Oware game created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwareEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/OwareNotAvailable"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/oware/{gameId}:
    get:
      operationId: getOwareGame
      tags: [Community]
      summary: Read a private Oware board
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/OwareGameId"
      responses:
        "200":
          description: Current privacy-safe board projection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwareEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/OwareNotAvailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/oware/{gameId}/moves:
    post:
      operationId: moveOwareGame
      tags: [Community]
      summary: Submit one revision-checked Oware move
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/OwareGameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwareMoveInput"
      responses:
        "200":
          description: Move accepted and current board returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwareEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/OwareNotAvailable"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    MemberBearer:
      type: http
      scheme: bearer
      bearerFormat: Opaque
      description: Short-lived member access token; the session subject must match memberId.
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Staff token carrying the explicitly documented capability and recent-MFA claim where required.
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Safe caller-provided identifier; invalid values are replaced.
      schema:
        $ref: "#/components/schemas/CorrelationId"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable key reused for retries of the same command.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
...
    CircleRoomId:
      name: circleId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    OwareGameId:
      name: gameId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  headers:
    CorrelationId:
      required: true
      description: Effective request correlation identifier.
      schema:
        $ref: "#/components/schemas/CorrelationId"
  responses:
    Unauthorized:
      description: A valid authenticated session is required.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    OwareNotAvailable:
      description: The game or its exact two-member circle context is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InvalidJSON:
      description: The body is malformed, oversized, contains unknown fields, or contains multiple values.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidJSON:
              value:
                error:
                  code: invalid_json
                  message: The request body must be one valid JSON object.
                meta:
                  correlationId: 01J9EXAMPLE000000000000000
    UnsupportedMediaType:
      description: Content-Type is not application/json.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationFailed:
      description: Request fields or the idempotency key are invalid.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidEmail:
              value:
                error:
                  code: validation_failed
                  message: One or more fields are invalid.
                  details:
                    - field: email
                      reason: must be a valid email address
                meta:
                  correlationId: 01J9EXAMPLE000000000000000
    InternalError:
      description: An unexpected server failure occurred.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    CorrelationId:
      type: string
      minLength: 8
      maxLength: 128
      pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
...
    Metadata:
      type: object
      additionalProperties: false
      required: [correlationId]
      properties:
        correlationId:
          $ref: "#/components/schemas/CorrelationId"
...
    OwareMoveInput:
      type: object
      additionalProperties: false
      required: [pit, expectedRevision]
      properties:
        pit:
          type: integer
          minimum: 0
          maximum: 11
        expectedRevision:
          type: integer
          format: int64
          minimum: 1
    OwareData:
      type: object
      additionalProperties: false
      required:
        [
          id,
          houses,
          captured,
          turn,
          yourPlayer,
          yourTurn,
          status,
          winner,
          revision,
          moveDeadline,
          serverTime,
        ]
      properties:
        id:
          type: string
        houses:
          type: array
          minItems: 12
          maxItems: 12
          items:
            type: integer
            minimum: 0
        captured:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: integer
            minimum: 0
        turn:
          type: string
          enum: [south, north]
        yourPlayer:
          type: string
          enum: [south, north]
        yourTurn:
          type: boolean
        status:
          type: string
          enum: [active, completed, expired]
        winner:
          type: integer
          minimum: -1
          maximum: 1
        revision:
          type: integer
          format: int64
          minimum: 1
        moveDeadline:
          type: string
          format: date-time
        serverTime:
          type: string
          format: date-time
    OwareEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/OwareData"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    MatchmakerProfileData:
      type: object
      additionalProperties: false
      required:
        [
          matchmakerId,
          displayName,
          licenseId,
          jurisdiction,
          licenseVersion,
          licenseValidUntil,
          minimumFeePesewas,
          maximumFeePesewas,
          languages,
          specialties,
          completedEngagements,
          ratingBasisPoints,
        ]
      properties:
        matchmakerId:
          type: string
        displayName:
          type: string
        licenseId:
          type: string
        jurisdiction:
          type: string
        licenseVersion:
          type: integer
          format: int64
          minimum: 1
        licenseValidUntil:
          type: string
          format: date-time
        minimumFeePesewas:
          type: integer
          format: int64
          minimum: 1
        maximumFeePesewas:
          type: integer
          format: int64
          minimum: 1
        languages:
          type: array
          items:
            type: string
        specialties:
          type: array
          items:
            type: string
        completedEngagements:
          type: integer
          format: int64
          minimum: 0
        ratingBasisPoints:
          type: integer
          minimum: 0
          maximum: 500
    MatchmakerProfileListData:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MatchmakerProfileData"
    MatchmakerProfileListEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/MatchmakerProfileListData"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    MatchmakerMilestoneInput:
      type: object
      additionalProperties: false
      required: [id, feePesewas, dueAfterDays]
      properties:
        id:
          type: string
        feePesewas:
          type: integer
          format: int64
          minimum: 1
        dueAfterDays:
          type: integer
          minimum: 0
          maximum: 365
    MatchmakerBookingInput:
      type: object
      additionalProperties: false
      required: [matchmakerId, termsId, termsVersion, milestones]
      properties:
        matchmakerId:
          type: string
        termsId:
          type: string
        termsVersion:
          type: integer
          format: int64
          minimum: 1
        milestones:
          type: array
          minItems: 1
          maxItems: 8
          items:
            $ref: "#/components/schemas/MatchmakerMilestoneInput"
    MatchmakerEngagementData:
      type: object
      additionalProperties: false
      required:
        [
          engagementId,
          matchmakerId,
          termsId,
          termsVersion,
          totalFeePesewas,
          bookedAt,
          memberConsented,
          candidateConsented,
          proposalExposed,
          completed,
          revision,
        ]
      properties:
        engagementId:
          type: string
        matchmakerId:
          type: string
        termsId:
          type: string
        termsVersion:
          type: integer
          format: int64
          minimum: 1
        totalFeePesewas:
          type: integer
          format: int64
          minimum: 1
        bookedAt:
          type: string
          format: date-time
        memberConsented:
          type: boolean
        candidateConsented:
          type: boolean
        proposalExposed:
          type: boolean
        completed:
          type: boolean
        proposalRef:
          type: string
        revision:
          type: integer
          format: int64
          minimum: 1
    MatchmakerEngagementEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/MatchmakerEngagementData"
        meta:
          $ref: "#/components/schemas/Metadata"
    MatchmakerEngagementListData:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MatchmakerEngagementData"
    MatchmakerEngagementListEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/MatchmakerEngagementListData"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          pattern: ^[a-z][a-z0-9_]*$
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        meta:
          $ref: "#/components/schemas/Metadata"
...